+++ .storybook/preview.tsx
@@ -10,10 +10,11 @@ import { ThemeProvider } from '../src/stylesheets';
 import GlobalStyle from '../src/stylesheets/global';
 
 export const parameters = {
+  viewMode: 'docs',
   options: {
     storySort: {
       method: 'alphabetical',
-      order: ['Brand-Story', 'Tokens', 'Components', 'Templates']
+      order: ['Brand Guidelines', 'Tokens', 'Components', 'Templates']
     }
   },
   viewport: {