+++ src/components/Textarea/Textarea.style.tsx
@@ -4,8 +4,7 @@ import { color, hideVisually, space } from '../../stylesheets';
 const StyledTextarea = styled.div<{ error?: boolean; success?: boolean; hideLabel?: boolean }>`
   label {
     display: block;
-    font-weight: bold;
-    color: ${color('neutral', 100)};
+    color: ${color('neutral')};
     margin-bottom: ${space('s')};
     ${props => props.hideLabel && hideVisually()}
   }
@@ -14,27 +13,15 @@ const StyledTextarea = styled.div<{ error?: boolean; success?: boolean; hideLabe
     width: 100%;
     outline: none;
     background: transparent;
-    color: ${color('neutral', 100)};
-    font-size: inherit;
-    font-weight: normal;
+    color: ${color('neutral')};
     resize: none;
     position: relative;
-    border: 1px solid ${color('neutral', 100)};
-    padding: ${space('m')};
+    border: 1px solid ${color('neutral')};
+    padding: ${space('s')};
 
     &::placeholder {
-      color: ${color('neutral', 100)};
-      font-weight: normal;
-    }
-
-    &::-ms-input-placeholder {
-      color: ${color('neutral', 100)};
-      font-weight: normal;
-    }
-
-    &:-ms-input-placeholder {
-      color: ${color('neutral', 100)};
-      font-weight: normal;
+      opacity: 1;
+      color: ${color('neutral', 30)};
     }
 
     &:focus {
@@ -45,6 +32,7 @@ const StyledTextarea = styled.div<{ error?: boolean; success?: boolean; hideLabe
   .textarea__error {
     display: block;
     color: ${color('warning')};
+    margin-top: ${space('xs')};
   }
 
   ${props =>